Настраивает сборку проекта
