+++ src/components/siteShowMoreButton.js
@@ -1,0 +1,7 @@
+const siteShowMoreButton = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {siteShowMoreButton};
