+++ src/components/siteTotalNumberFilms.js
@@ -1,0 +1,7 @@
+const siteTotalNumberFilms = () => {
+  return (
+    `<p>130 291 movies inside</p>`
+  );
+};
+
+export {siteTotalNumberFilms};
